replace `Event.getResult()` and `.isCancelable()` transformer with ClassValue
