+++ src/main/java/net/minecraftforge/fml/relauncher/FMLCorePlugin.java
@@ -30,7 +30,6 @@ public class FMLCorePlugin implements IFMLLoadingPlugin
     {
         return new String[] {
                              "net.minecraftforge.fml.common.asm.transformers.SideTransformer",
-                             "net.minecraftforge.fml.common.asm.transformers.EventSubscriptionTransformer",
                              "net.minecraftforge.fml.common.asm.transformers.EventSubscriberTransformer",
                              "net.minecraftforge.fml.common.asm.transformers.SoundEngineFixTransformer",
                              "net.minecraftforge.fml.common.asm.transformers.LWJGLTransformer",
